refactor(lsp): split module specifier helpers
